Break down db module into categories
